Automate database setup
Add a migration system for applying indexes or otherwise making changes
to the MongoDB.

A new package.json script is now added called `migrate-db`, which can be
used to migrate the database up to the current version or down to
previous versions.

```
$ yarn migrate-db up
$ yarn migrate-db down 0
```

Adding logic to the application to automatically migrate the database on
application startup will be added in a future change.