Finish transition to env-based config
Remove the previous fixed environment config files from the core code
and move them into the scripts directory. The old object-based configs
are now purely there to facilitate the set-config script, which itself
may eventually be removed or turned into a script for converting JSON to
.env or vice versa.

Added a mock for the config module for tests. (It may be worth exploring
using dotenv in the test setup.)