GH-28: Redo character list as left-side pills
Use custom tab layout based on a column of character name buttons to
select the desired character tab.

This is in preparation for retrieving character IDs from the API and
being able to dynamically grow the character name column/retrieve from
the database in pages rather than all at once.