GH-28: Support limit and cursor parameters on character list
Add limit and cursor parameters to character list REST API. limit
specifies how many results to return in one request, and cursor is an
opaque (to the caller) field that can be used to start the search
partway. The cursor value is supplied by the response from the character
list request in the nextCursor field (which will be null if there are no
more results).