GH-37: Update character list component
Character list component will now trigger the initial load of character
IDs when rendered (assuming that characters.pending is true, which
should only be the case when the characters store is reset).

Includes loading spinner to display while loading is in progress and a
"more" button that can be clicked to return additional pages of
character data.