GH-43: Add API actions and reducers for PATCH and DELETE
Add API actions that will invoke the per-character PATCH and DELETE
verbs, and the corresponding reducers that will update state
accordingly.

The PATCH action will set `{ patchStatus: { loading: true } }` on the
character being patched when the request begins, update the character
according to the patch value when the request completes, and set an
error object inside `patchStatus` when the request fails.

The DELETE action will set `{ deleteStatus: { loading: true } }` on the
character being deleted when the request begins, remove the character
from the store when the request completes, and set an error object
inside `deleteStatus` when the request fails. On successful deletion,
the `selectedId` prop will be set to the first remaining character, or
left empty if the last character is deleted.