GH-43: Add support for deleting characters
Adds a UI path for deleting a character.

The UI is a delete button on the header bar of the selected
character; when clicked it pops up a modal with cancel and confirm
buttons. The cancel button will dismiss the modal; the confirm button
dismisses the modal and replaces the delete button with a spinner. Once
the delete completes, the spinner (and the deleted character) disappers.

Also includes a minor tweak to the async mode for ClickableEdit to make
the spinners look better on the default dark background of the edit
control.